+++ package.json
@@ -7,7 +7,7 @@
   "license": "UNLICENSED",
   "scripts": {
     "typeorm": "typeorm-ts-node-commonjs",
-    "migration:run": "npm run build && npm run typeorm migration:run -- -d ./src/migrations-data-source.ts",
+    "migration:run": "npm run typeorm migration:run -- -d ./src/migrations-data-source.ts",
     "prebuild": "rimraf dist",
     "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
