+++ src/modules/albums/albums.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AlbumsService } from './albums.service';
 import { AlbumsController } from './albums.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Album } from './entities/album.entity.js';
+import { Album } from './entities/album.entity';
 
 @Module({
   controllers: [AlbumsController],
