+++ src/modules/albums/dto/update-album.dto.ts
