+++ src/modules/albums/entities/album.entity.ts
@@ -1,4 +1,4 @@
-import { Artist } from '../../artists/entities/artist.entity.js';
+import { Artist } from '../../artists/entities/artist.entity';
 import {
   Column,
   Entity,
