+++ src/modules/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ArtistsModule } from './artists/artists.module.js';
+import { ArtistsModule } from './artists/artists.module';
 import { UsersModule } from './users/users.module';
 import { AlbumsModule } from './albums/albums.module';
 import { TracksModule } from './tracks/tracks.module';
