+++ src/modules/artists/artists.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ArtistsService } from './artists.service';
 import { ArtistsController } from './artists.controller';
-import { Artist } from './entities/artist.entity.js';
+import { Artist } from './entities/artist.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 @Module({
