+++ src/modules/artists/dto/create-artist.dto.ts
