+++ src/modules/artists/entities/artist.entity.ts
