+++ src/modules/favourites/favourites.module.ts
@@ -1,15 +1,15 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { FavouritesService } from './favourites.service';
 import { FavouritesController } from './favourites.controller';
-import { TracksModule } from './../tracks/tracks.module.js';
-import { ArtistsModule } from './../artists/artists.module.js';
-import { AlbumsModule } from './../albums/albums.module.js';
+import { TracksModule } from '../tracks/tracks.module';
+import { ArtistsModule } from './../artists/artists.module';
+import { AlbumsModule } from './../albums/albums.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import {
   AlbumFavorites,
   ArtistFavorites,
   TrackFavorites,
-} from './entities/favourites.entity.js';
+} from './entities/favourites.entity';
 
 @Module({
   controllers: [FavouritesController],
