+++ src/modules/tracks/tracks.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TracksService } from './tracks.service';
 import { TracksController } from './tracks.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Track } from './entities/track.entity.js';
+import { Track } from './entities/track.entity';
 
 @Module({
   controllers: [TracksController],
