+++ src/modules/tracks/tracks.service.ts
@@ -8,7 +8,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
-import { Track } from './entities/track.entity.js';
+import { Track } from './entities/track.entity';
 
 @Injectable()
 export class TracksService {
