+++ src/modules/users/users.controller.ts
@@ -15,7 +15,7 @@ import {
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { User } from './entities/user.entity.js';
+import { User } from './entities/user.entity';
 
 @Controller('user')
 export class UsersController {
