+++ src/modules/users/users.service.ts
@@ -7,7 +7,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { User } from './entities/user.entity.js';
+import { User } from './entities/user.entity';
 
 @Injectable()
 export class UsersService {
